backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
supabase
python-multipart
google-generativeai
langgraph
openai
psycopg2-binary
docx2txt
fitz
py-mupdf
